Rename packages/keystone/src/types/generated.ts to packages/keystone/src/types/type-info.ts
