+++ packages/keystone/src/types/index.ts
@@ -7,4 +7,4 @@ export * from './context';
 export * from './next-fields';
 export * as filters from './filters';
 export { jsonFieldTypePolyfilledForSQLite } from './json-field-type-polyfill-for-sqlite';
-export * from './generated';
+export * from './type-info';
